pyproject.toml
[project]
name = "wavewhiz"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "anapauladominoni", email = "anapauladom29@gmail.com"},
]
dependencies = ["django>=5.2.6", "djangorestframework>=3.16.1", "django-cors-headers>=4.7.0", "pillow>=11.3.0", "djangorestframework-simplejwt>=5.5.1"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
packages = ["wavewhiz_app"]

[tool.pdm.scripts]
runserver = "python manage.py runserver"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
createsuperuser = "python manage.py createsuperuser"